maskedimage moved out of nextjs app
