+++ src/app/components/MaskedImage/index.test.tsx
