+++ src/app/components/MaskedImage/styles.tsx
